refactor(tui): move runtime adapters to named owners
