+++ station/src/client/fixtures/mockObserverSnapshot.ts
