+++ station/src/station/store/stationCommandDispatch.test.ts
@@ -11,8 +11,8 @@ import {
 } from "@station/dashboard-core/runtime";
 import type { DashboardCapabilities } from "@station/dashboard-core/runtime";
 import { dashboardRowIds, selectDashboardSlots } from "@station/dashboard-core/selectors";
-import { createObserverStationClient } from "../../sources/observerStationClient.js";
-import type { StationClient } from "../../sources/types.js";
+import { createObserverStationClient } from "../../client/observerStationClient.js";
+import type { StationClient } from "../../client/stationClient.js";
 import { waitFor } from "../../terminal/testing/waitFor.js";
 import { createFakeFolderService } from "../test/support/fakeFolderService.js";
 import type { StationMouseEvent } from "../../input/mouse.js";
